[wip] refactor of static map into pathwidget
