+++ client/src/scss/pathWidget.scss
@@ -1,0 +1,3 @@
+.mapWpr {
+  margin-right: 0;
+}
